Move get_subsystem from System to Group
