+++ openmdao/core/system.py
@@ -8,7 +8,6 @@ from collections import namedtuple, OrderedDict
 import numpy
 
 from six.moves import range
-from six import string_types
 
 from openmdao.proc_allocators.default_allocator import DefaultAllocator
 from openmdao.jacobians.default_jacobian import DefaultJacobian
@@ -880,38 +879,6 @@ class System(object):
     def _linearize(self):
         """Compute jacobian / factorization."""
         pass
-
-    def get_subsystem(self, name):
-        """Return the system called 'name' in the current namespace.
-
-        Args
-        ----
-        name : str
-            name of the desired system in the current namespace.
-
-        Returns
-        -------
-        System or None
-            System if found else None.
-        """
-        idot = name.find('.')
-
-        # If name does not contain '.', only check the immediate children
-        if idot == -1:
-            for subsys in self._subsystems_allprocs:
-                if subsys.name == name:
-                    return subsys
-        # If name does contain at least one '.', we have to recurse (possibly).
-        else:
-            sub_name = name[:idot]
-            for subsys in self._subsystems_allprocs:
-                # We only check if the prefix matches, and with the prefix removed.
-                if subsys.name == sub_name:
-                    result = subsys.get_subsystem(name[idot + 1:])
-                    if result:
-                        return result
-
-        return None
 
     def initialize(self):
         """Optional user-defined method run once during instantiation.
